+++ packages/terra-graphs-docs/src/terra-dev-site/test/graphs/GanttChart/Reflow/DynamicData.test.jsx
@@ -1,23 +1,33 @@
 import React from 'react';
-import GanttChart from '@cerner/terra-graphs/lib/components/Gantt/GanttChart';
+import Carbon from '@cerner/carbon-graphs/lib/js/carbon';
 import Button from 'terra-button/lib/Button';
 import utils from '@cerner/carbon-graphs/lib/js/helpers/utils';
 import '@cerner/terra-graphs-docs/lib/Css/ExampleGraphContainer.module.scss';
+import '@cerner/terra-graphs/lib/components/Graph.module.scss';
+import '@cerner/terra-graphs/lib/components/Gantt/GanttChart.module.scss';
 import getGanttPanningConfig from '@cerner/terra-graphs-docs/lib/example-datasets/graphConfigObjects/Gantt/ganttPanning';
-import data from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/dynamicPanningData';
+import exampleData from '@cerner/terra-graphs-docs/lib/example-datasets/dataObjects/Gantt/dynamicPanningData';
+
+const graphConfig = utils.deepClone(getGanttPanningConfig('#GanttPanningDynamicData'));
+const dataset = [utils.deepClone(exampleData[0])];
 
 const initialState = {
   initial: 0,
   factor: 3,
-  dataset: [utils.deepClone(data[0])],
-  graphConfig: utils.deepClone(getGanttPanningConfig('#GanttPanningDynamicData')),
 };
 
+let graph;
+
 const GanttPanningExample = () => {
+  React.useEffect(() => {
+    graph = Carbon.api.gantt(graphConfig);
+    dataset.forEach((data) => {
+      graph.loadContent(data);
+    });
+  }, []);
+
   const reducer = (panState, action) => {
-    const newGraphState = utils.deepClone(panState.graphConfig);
     let hour;
-    const newDataset = [utils.deepClone(data[1])];
 
     switch (action.type) {
       case 'panLeft':
@@ -30,25 +40,29 @@ const GanttPanningExample = () => {
         return panState;
     }
 
-    newGraphState.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
-    newGraphState.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+    graph.config.axis.x.lowerLimit = new Date(2016, 0, 1, hour).toISOString();
+    graph.config.axis.x.upperLimit = new Date(2016, 0, 2, hour).toISOString();
+
+    const newDataset = {
+      panData: [utils.deepClone(utils.deepClone(exampleData[1]))],
+    };
+
+    graph.reflowMultipleDatasets(newDataset);
 
     return {
       initial: hour,
       factor: panState.factor,
-      dataset: { panData: newDataset },
-      graphConfig: utils.deepClone(newGraphState),
     };
   };
 
-  const [panState, dispatch] = React.useReducer(reducer, initialState);
+  const [, dispatch] = React.useReducer(reducer, initialState);
 
   return (
     <React.Fragment>
       <Button className="button-pan-left" text="<" onClick={() => dispatch({ type: 'panLeft' })} />
       <Button className="button-pan-right" text=">" onClick={() => dispatch({ type: 'panRight' })} />
       <div id="tooltip" className="initial-tooltip" />
-      <GanttChart graphID="GanttPanningDynamicData" graphConfig={panState.graphConfig} dataset={panState.dataset} />
+      <div id="GanttPanningDynamicData" />
     </React.Fragment>
   );
 };